HOCS-2274: Move Parse and Format Date to date helpers.
To apply the DRY principle, the date parsing and formatting has moved to be part of the dateHelpers file. This also aids the single responsibility principle.
