+++ server/libs/dateHelpers.js
@@ -12,4 +12,14 @@ const getUtcDateString = (date) => {
     return undefined;
 };
 
-module.exports = { addDays, getUtcDateString };
+const parseDate = (rawDate) => {
+    const [date] = rawDate.match(/\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])/g) || [];
+    if (!date) {
+        return null;
+    }
+    const [year, month, day] = date.split('-');
+    return `${day}/${month}/${year}`;
+};
+const formatDate = (date) => date ? parseDate(date) : null;
+
+module.exports = { addDays, getUtcDateString, formatDate };
