+++ server/lists/adapters/workstacks.js
@@ -1,4 +1,4 @@
-const { addDays } = require('../../libs/dateHelpers');
+const { addDays, formatDate } = require('../../libs/dateHelpers');
 
 const byUser = (userId) => ({ userUUID }) => userUUID === userId;
 const byCaseReference = (a, b) => a.caseReference.localeCompare(b.caseReference);
@@ -108,16 +108,6 @@ const returnMyCasesWorkstackColumns = (configuration, workstackData) => {
 
     return getColumnsForMyCases.workstackColumns;
 };
-
-const parseDate = (rawDate) => {
-    const [date] = rawDate.match(/[0-9]{4}-[0-1][0-9]-[0-3][0-9]/g) || [];
-    if (!date) {
-        return null;
-    }
-    const [year, month, day] = date.split('-');
-    return `${day}/${month}/${year}`;
-};
-const formatDate = (date) => date ? parseDate(date) : null;
 
 const bindDisplayElements = fromStaticList => async (stage) => {
     stage.assignedTeamDisplay = await fromStaticList('S_TEAMS', stage.teamUUID);
